# server/src/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "UniVote API Documentation",
        "description": "REST API specification for the UniVote ACSES UMaT E-Voting System. Administrative actions require bearer tokens.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Base Server"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "summary": "Query all user registry profiles",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User registry list (masked details)."
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "summary": "Inspect individual unmasked user",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unmasked user details. Strictly audit-logged."
                    }
                }
            },
            "patch": {
                "summary": "Update user details",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "studentId": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "voter",
                                            "admin",
                                            "auditor"
                                        ]
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "phoneNumber": {
                                        "type": "string"
                                    },
                                    "year": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registry profile updated successfully."
                    }
                }
            },
            "delete": {
                "summary": "Delete user profile",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
...
    }
}